retizero-service/RetiZero/retinal-thesis/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
albumentations>=1.3.1
wandb>=0.16.0
scikit-learn>=1.3.0
scikit-multilearn>=0.2.0
iterstrat>=0.1.7
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
pytorch-grad-cam>=1.5.0
netcal>=1.3.5
